Again have to cast to qulonglong... weird...
git-svn-id: svn+ssh://svn.code.sf.net/p/yabause/code/trunk@3155 d7be59e3-070e-4b9a-85a1-4646befae34d

+++ yabause/src/qt/ui/UISettings.cpp
@@ -151,7 +151,7 @@ void UISettings::getCdDriveList()
 		if (sscanf(buffer, "drive name:%s", drive_name) == 1) {
 			sprintf(drive_path, "/dev/%s", drive_name);
 
-			cbCdDrive->addItem(drive_path, NULL);
+			cbCdDrive->addItem(drive_path, (qulonglong) NULL);
 		}
 	}
 #elif defined Q_OS_MAC
